Reworked TokenInfoPopover so it handles token data not being availabled
